% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_plots.R
\name{comparison_plot}
\alias{comparison_plot}
\title{RnBeads}
\usage{
comparison_plot(wd, f)
}
\arguments{
\item{wd}{path to the rnbeads results directory where csv files are located and name of csv file}
}
\value{
list of p values
}
\description{
Takes in a working directory and csv file name and returns a list of differential methylation p values from the comparisons csv files located in differential_methylation_data dir of Rnbeads Analysis
}

