% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_datasets.R
\name{datasets_total}
\alias{datasets_total}
\title{All Datasets List used in RnBeads Analysis}
\usage{
datasets_total(rd)
}
\arguments{
\item{rd}{A path to the rnbeads results repository}
}
\value{
list of directories
}
\description{
Takes in a results directory and returns a list of grouped directories that used the same set of datasets
}

