% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_perfomed.R
\name{modules_performed}
\alias{modules_performed}
\title{RnBeads modules performed}
\usage{
modules_performed(wd)
}
\arguments{
\item{wd}{path to the rnbeads results directory where analysis.log file is located}
}
\value{
list of rnbeads modules performed
}
\description{
Takes in a working directory and returns a list of rnbeads modules performed
}

