% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_scorer.R
\name{multipleVennDiagram}
\alias{multipleVennDiagram}
\title{Draw a Venn Diagram}
\usage{
multipleVennDiagram(cb, results.dir, nrows.value)
}
\arguments{
\item{x}{A numeric value to be squared}
}
\value{
Squared value of x
}
\description{
Takes in any numeric value and square it
}

