% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comparison_data.R
\name{readComparisonData}
\alias{readComparisonData}
\title{Get Comparison file data}
\usage{
readComparisonData(qq.value, qq.dir, comp.index, topRows)
}
\arguments{
\item{x}{A numeric value to be squared}
}
\value{
Squared value of x
}
\description{
Takes in any numeric value and square it
}

