% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_type.R
\name{rnbi.analysis.datatype}
\alias{rnbi.analysis.datatype}
\title{RnBeads analysis type (IDAT, RRBS etc..)}
\usage{
rnbi.analysis.datatype(wd, analysis.name)
}
\arguments{
\item{wd}{path to the rnbeads results directory and analysis.name is the selected analysis}

\item{analysis.name}{of the RnBeads generated report (root directory name)}
}
\value{
text containg idat identifier to check and declare idat analysis or not
}
\description{
rnbi.analysis.datatype
}
\details{
return the analysis data type of a specified analysis from the table in data_import.html file

Takes in a working directory and analysis name and return the analysis data type of a specified analysis from the table in data_import.html file
}
