% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_perfomed.R
\name{rnbi.analysis.rrbs.modules.performed}
\alias{rnbi.analysis.rrbs.modules.performed}
\title{rnbi.analysis.rrbs.modules.performed}
\usage{
rnbi.analysis.rrbs.modules.performed(wd)
}
\arguments{
\item{wd}{path to the rnbeads results directory where index.html file is located}
}
\value{
list of rnbeads modules performed
}
\description{
Takes in a working directory and returns a list of rnbeads modules performed for RRBS analysis
}

