% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{rnbi.dataset}
\alias{rnbi.dataset}
\title{All Datasets List used in RnBeads Analysis}
\usage{
rnbi.dataset(rd)
}
\arguments{
\item{rd}{A path to the rnbeads analysis folder}
}
\value{
list of common and uncommon datasets index
}
\description{
rnbi.dataset
}
\details{
return the list of datasets/sample sheet used in the RnBeads analysis found in the given path

Takes in a results directory and returns the list of total common and uncommon datasets/sample sheet used in the RnBeads analysis found in the given path
}

