% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot.R
\name{rnbi.qqplot.data.single}
\alias{rnbi.qqplot.data.single}
\title{rnbi.qqplot.data.single
For Idat analysis
convert the p-values of a one analysis in to an object of class qqplot.data which will be used to draw qqplot .}
\usage{
rnbi.qqplot.data.single(x, p = "P")
}
\arguments{
\item{x}{a dataframe of differential methlytion p-value}

\item{p}{column consist of differential methylation p-values}
}
\value{
return an object of class qqplot.data which will be used to draw qqplot .
}
\description{
rnbi.qqplot.data.single
For Idat analysis
convert the p-values of a one analysis in to an object of class qqplot.data which will be used to draw qqplot .
}
