% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot.R
\name{rnbi.qqplot.data.static.rrbs}
\alias{rnbi.qqplot.data.static.rrbs}
\title{rnbi.qqplot.data.static
 For RRBS analysis
read and convert the p values data into a list so that it can be used to in the qqplot function "qqunif.plot"}
\usage{
rnbi.qqplot.data.static.rrbs(wd, f)
}
\arguments{
\item{wd}{a working directory of RnBeads reports}

\item{f}{a index of the comparison file}
}
\value{
return an object that can be used to in the qqplot function "qqunif.plot"
}
\description{
rnbi.qqplot.data.static
 For RRBS analysis
read and convert the p values data into a list so that it can be used to in the qqplot function "qqunif.plot"
}
