% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot.R
\name{rnbi.qqplot.double}
\alias{rnbi.qqplot.double}
\title{rnbi.qqplot.double}
\usage{
rnbi.qqplot.double(x, y)
}
\arguments{
\item{x}{a dataframe of differential methlytion p-value from one analysis}

\item{y}{a dataframe of differential methlytion p-value from second analysis}
}
\value{
return an object to draw qqplot for two analysis simultaneously using plotly
}
\description{
draw qqplot for two analysis
}

