% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot.R
\name{rnbi.qqplot.double.rrbs}
\alias{rnbi.qqplot.double.rrbs}
\title{rnbi.qqplot.double.rrbs
 For RRBS analysis
draw qqplot for two analysis}
\usage{
rnbi.qqplot.double.rrbs(x, y)
}
\arguments{
\item{x}{a dataframe of differential methlytion p-value from one analysis}

\item{y}{a dataframe of differential methlytion p-value from second analysis}
}
\value{
return an object to draw qqplot for two analysis simultaneously using plotly for RRBS analysis
}
\description{
rnbi.qqplot.double.rrbs
 For RRBS analysis
draw qqplot for two analysis
}
