% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot.R
\name{rnbi.qqplot.single}
\alias{rnbi.qqplot.single}
\title{rnbi.qqplot.single}
\usage{
rnbi.qqplot.single(x)
}
\arguments{
\item{x}{a dataframe of differential methlytion p-value}
}
\value{
return a plot object which can be used to draw qqplot of p-values using plotly
}
\description{
Draw the qqplot for single data for idat analysis
Takes in a differential methlytion p-value data frame and return a plot object which can be used to draw qqplot of p-values using plotly
}
