% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot.R
\name{rnbi.qqplot.single.rrbs}
\alias{rnbi.qqplot.single.rrbs}
\title{rnbi.qqplot.single.rrbs}
\usage{
rnbi.qqplot.single.rrbs(x)
}
\arguments{
\item{x}{a dataframe of differential methlytion p-value}
}
\value{
return an object to draw the qqplot for single data for RRBS analysis
}
\description{
Draw the qqplot for single data for RRBS analysis
}
