% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comparison_data.R
\name{rnbi.read.comparisondata.rrbs}
\alias{rnbi.read.comparisondata.rrbs}
\title{rnbi.read.comparisondata.rrbs
Reads the rrbs comparison csv file under differential_methylation_data folder of the selected analysis and returns the results.}
\usage{
rnbi.read.comparisondata.rrbs(qq.value, qq.dir, comp.index, topRows,
  columnSelected, results.dir)
}
\arguments{
\item{qq.value}{The analysis name}

\item{qq.dir}{analysis path}

\item{comp.index}{the comparison file index (e.g. 1,2,3..)}

\item{topRows}{reading top n rows}

\item{columnSelected}{which column to read the data}

\item{results.dir}{absolute path to the analysis folder}
}
\value{
dataframe of the data read from csv file
}
\description{
rnbi.read.comparisondata.rrbs
Reads the rrbs comparison csv file under differential_methylation_data folder of the selected analysis and returns the results.
}
