% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comparison_data.R
\name{rnbi.read.comparisondatalogical}
\alias{rnbi.read.comparisondatalogical}
\title{rnbi.read.comparisondatalogical}
\usage{
rnbi.read.comparisondatalogical(dataset, column, equality, range)
}
\arguments{
\item{dataset}{dataframe of differential methylation comparison data,}

\item{column}{column selected from the interface to read the data}

\item{equality}{the assignment operator selected form the interface}

\item{range}{range of value selected from the interface}
}
\value{
dataframe of the data read from csv file
}
\description{
Reads the comparison csv file under differential_methylation_data folder of the selected analysis and returns the Logical (True, False) results used for Venn Diagram purpose.
}

