% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{rnbi.total.analysis}
\alias{rnbi.total.analysis}
\title{rnbi.total.analysis
return the list of RnBeads analysis found in the given path}
\usage{
rnbi.total.analysis(wd)
}
\arguments{
\item{wd}{A path to the rnbeads analysis folder}
}
\value{
list of analysis found
}
\description{
Takes in a results directory and returns a list of analysis found in that directories
}
