% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApplication.R
\name{runApplication}
\alias{runApplication}
\title{Launch ShinyBeads app}
\usage{
runApplication(wd)
}
\arguments{
\item{wd}{A path to the rnbeads analysis folder}
}
\description{
Takes in a results directory and Run the ShinyBeads application
}
