% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot.R
\name{qqunif.plot}
\alias{qqunif.plot}
\title{qqunif.plot}
\usage{
qqunif.plot(pvalues, should.thin = T, thin.obs.places = 2,
  thin.exp.places = 2, xlab = expression(paste("Expected (", -log[10],
  " p-value)")), ylab = expression(paste("Observed (", -log[10],
  " p-value)")), draw.conf = TRUE, conf.points = 1000,
  conf.col = "lightgray", conf.alpha = 0.05, already.transformed = FALSE,
  pch = 20, aspect = "fill", prepanel = prepanel.qqunif,
  par.settings = list(superpose.symbol = list(pch = pch)))
}
\arguments{
\item{pvalues}{p-values}

\item{should.thin}{thining}

\item{thin.obs.places}{thining observed}

\item{thin.exp.places}{thining expected}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{draw.conf}{draw the conf}

\item{conf.points}{points}

\item{conf.col}{columns}

\item{conf.alpha}{alpha}

\item{already.transformed}{transformation}

\item{pch}{pch}

\item{aspect}{aspect}

\item{prepanel}{prepanel}

\item{par.settings}{settings}
}
\value{
functiont to draw qqplot
}
\description{
Takes in a p-values
}
